Adding wallet money activity completed Delete multiple cart items task added Update customer wallet task added Customer profile activity created and completed Payment success activity completed customer profile layout xml created

+++ app/src/main/java/com/theleafapps/shopnick/models/Customer.java
@@ -34,7 +34,7 @@ public class Customer extends BaseRecord{
     public String country;
 
     @JsonProperty("wallet_value")
-    public Float wallet_value;
+    public double wallet_value;
 
     @JsonProperty("zipcode")
     public String zipcode;
